# span: fix corruption bug in btree span frontier