--- a/pkg/util/span/frontier.go
+++ b/pkg/util/span/frontier.go
@@ -104,7 +104,7 @@ func (r OpResult) asBool() bool {
 type Operation func(roachpb.Span, hlc.Timestamp) (done OpResult)
 
 var useBtreeFrontier = envutil.EnvOrDefaultBool("COCKROACH_BTREE_SPAN_FRONTIER_ENABLED",
-	metamorphic.ConstantWithTestBool("COCKROACH_BTREE_SPAN_FRONTIER_ENABLED", false))
+	metamorphic.ConstantWithTestBool("COCKROACH_BTREE_SPAN_FRONTIER_ENABLED", true))
 
 func enableBtreeFrontier(enabled bool) func() {
 	old := useBtreeFrontier
@@ -375,6 +375,12 @@ func (f *btreeFrontier) mergeEntries(e *btreeFrontierEntry) (*btreeFrontierEntry
 		f.mergeAlloc = append(f.mergeAlloc, rightIter.Cur())
 	}
 
+	// If there were no left or right merges, return without restructuring the
+	// tree.
+	if len(f.mergeAlloc) == 0 {
+		return leftMost, nil
+	}
+
 	// Delete entries first, before updating leftMost boundaries since doing so
 	// will mess up btree.
 	for i, toRemove := range f.mergeAlloc {
@@ -384,10 +390,8 @@ func (f *btreeFrontier) mergeEntries(e *btreeFrontierEntry) (*btreeFrontierEntry
 		}
 	}
 
-	leftMost.End = end
-	if expensiveChecksEnabled() {
-		leftMost.spanCopy.EndKey = append(roachpb.Key{}, end...)
-	}
+	f.setEndKey(leftMost, end)
+
 	return leftMost, nil
 }
 
@@ -435,17 +439,29 @@ func (f *btreeFrontier) splitEntryAt(
 
 	right = newFrontierEntry(&f.idAlloc, split, e.End, e.ts)
 
-	// Adjust e boundary before we add right (so that there is no overlap in the tree).
-	e.End = split
-	if expensiveChecksEnabled() {
-		e.spanCopy.EndKey = append(roachpb.Key{}, split...)
-	}
+	// Adjust e boundary before we add right (so that there is no overlap in the
+	// tree).
+	f.setEndKey(e, split)
 
 	if err := f.setEntry(right); err != nil {
 		putFrontierEntry(right)
 		return nil, nil, err
 	}
 	return e, right, nil
+}
+
+// setEndKey changes the end key assigned to the entry. setEndKey requires the
+// entry to be in the tree.
+func (f *btreeFrontier) setEndKey(e *btreeFrontierEntry, endKey roachpb.Key) {
+	// The tree implementation expects the Start and End keys of a span to be
+	// immutable. We remove the leftMost node before updating the End index in
+	// order to avoid corrupting the `maxKey` that is inlined in the `node`.
+	f.tree.Delete(e)
+	e.End = endKey
+	if expensiveChecksEnabled() {
+		e.spanCopy.EndKey = append(roachpb.Key{}, endKey...)
+	}
+	f.tree.Set(e)
 }
 
 // forward is the work horse of the btreeFrontier.  It forwards the timestamp