Previously, the btree span frontier implementation modified the EndKey of btree entries. The btree internally expects the EndKey to be immutable, so it caused queries for containing spans to return incorrect results.

Now, entries are removed from the btree before their end key is modified. This is expected to have minimal impact on performance because the fast/common path does not require merging or splitting entries.

Fixes: #120987