@jeffswenson 2024-10-01
```
➜  go test . --bench=BenchmarkFrontier
goos: darwin
goarch: arm64
pkg: github.com/cockroachdb/cockroach/pkg/util/span
BenchmarkFrontier/btree=false/rnd-10              834922              1341 ns/op             691 B/op         16 allocs/op
BenchmarkFrontier/btree=false/r=128-10             44199             26000 ns/op            5565 B/op        101 allocs/op
BenchmarkFrontier/btree=false/r=1024-10             4670            249546 ns/op           42526 B/op        721 allocs/op
BenchmarkFrontier/btree=false/r=4096-10             1012           1150424 ns/op          174038 B/op       2852 allocs/op
BenchmarkFrontier/btree=false/r=8192-10              495           2371999 ns/op          349876 B/op       5615 allocs/op
BenchmarkFrontier/btree=false/r=16384-10             236           5053206 ns/op          715567 B/op      11252 allocs/op
BenchmarkFrontier/btree=true/rnd-10              2163332               554.5 ns/op             0 B/op          0 allocs/op
BenchmarkFrontier/btree=true/r=128-10              99777             11327 ns/op               0 B/op          0 allocs/op
BenchmarkFrontier/btree=true/r=1024-10              9013            127850 ns/op               0 B/op          0 allocs/op
BenchmarkFrontier/btree=true/r=4096-10              1990            569113 ns/op               0 B/op          0 allocs/op
BenchmarkFrontier/btree=true/r=8192-10               826           1410739 ns/op               0 B/op          0 allocs/op
BenchmarkFrontier/btree=true/r=16384-10              433           2643127 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/cockroachdb/cockroach/pkg/util/span  41.653s

➜  git checkout jeffswenson-fix-btree-span-frontier
➜  go test . --bench=BenchmarkFrontier
goos: darwin
goarch: arm64
pkg: github.com/cockroachdb/cockroach/pkg/util/span
BenchmarkFrontier/btree=false/rnd-10              842134              1349 ns/op             693 B/op         16 allocs/op
BenchmarkFrontier/btree=false/r=128-10             47208             26647 ns/op            5688 B/op        103 allocs/op
BenchmarkFrontier/btree=false/r=1024-10             4950            250329 ns/op           42245 B/op        715 allocs/op
BenchmarkFrontier/btree=false/r=4096-10             1072           1146503 ns/op          169264 B/op       2786 allocs/op
BenchmarkFrontier/btree=false/r=8192-10              523           2319801 ns/op          338960 B/op       5466 allocs/op
BenchmarkFrontier/btree=false/r=16384-10             242           4700627 ns/op          653660 B/op      10352 allocs/op
BenchmarkFrontier/btree=true/rnd-10              1850320               648.5 ns/op             0 B/op          0 allocs/op
BenchmarkFrontier/btree=true/r=128-10             103184             11768 ns/op               0 B/op          0 allocs/op
BenchmarkFrontier/btree=true/r=1024-10              9547            127218 ns/op               0 B/op          0 allocs/op
BenchmarkFrontier/btree=true/r=4096-10              2056            591642 ns/op               0 B/op          0 allocs/op
BenchmarkFrontier/btree=true/r=8192-10               858           1403478 ns/op               0 B/op          0 allocs/op
BenchmarkFrontier/btree=true/r=16384-10              445           2640383 ns/op               0 B/op          0 allocs/op
```