@jeffswenson 2024-10-02
```
[nix-shell:~/Code/cockroach]$ benchstat before.txt after.txt
goos: darwin
goarch: arm64
                            │ before.txt  │              after.txt              │
                            │   sec/op    │   sec/op     vs base                │
Frontier/btree=true/rnd       647.9n ± 3%   757.4n ± 1%  +16.91% (p=0.000 n=10)
Frontier/btree=true/r=128     13.42µ ± 3%   13.53µ ± 5%        ~ (p=0.912 n=10)
Frontier/btree=true/r=1024    154.3µ ± 1%   151.6µ ± 2%   -1.74% (p=0.004 n=10)
Frontier/btree=true/r=4096    704.1µ ± 1%   704.8µ ± 4%        ~ (p=0.579 n=10)
Frontier/btree=true/r=8192    1.742m ± 1%   1.684m ± 3%   -3.34% (p=0.001 n=10)
Frontier/btree=true/r=16384   3.305m ± 1%   3.263m ± 4%        ~ (p=0.165 n=10)
geomean                       132.6µ        134.9µ        +1.69%

                            │  before.txt  │              after.txt              │
                            │     B/op     │    B/op     vs base                 │
Frontier/btree=true/rnd       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Frontier/btree=true/r=128     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Frontier/btree=true/r=1024    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Frontier/btree=true/r=4096    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Frontier/btree=true/r=8192    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Frontier/btree=true/r=16384   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                  ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                            │  before.txt  │              after.txt              │
                            │  allocs/op   │ allocs/op   vs base                 │
Frontier/btree=true/rnd       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Frontier/btree=true/r=128     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Frontier/btree=true/r=1024    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Frontier/btree=true/r=4096    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Frontier/btree=true/r=8192    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Frontier/btree=true/r=16384   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                  ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```